+++ apps/etterlatte-behandling/src/main/kotlin/behandling/aktivitetsplikt/AktivitetspliktRoute.kt
@@ -12,7 +12,7 @@ import io.ktor.server.routing.post
 import io.ktor.server.routing.route
 import io.ktor.util.pipeline.PipelineContext
 import kotlinx.coroutines.runBlocking
-import no.nav.etterlatte.behandling.aktivitetsplikt.vurdering.LagreAktivitetspliktAktivitetsgrad
+import no.nav.etterlatte.behandling.aktivitetsplikt.vurdering.LagreAktivitetspliktAktivitetsgradMedUnntak
 import no.nav.etterlatte.behandling.aktivitetsplikt.vurdering.LagreAktivitetspliktUnntak
 import no.nav.etterlatte.behandling.domain.TilstandException
 import no.nav.etterlatte.brev.model.BrevID
@@ -301,7 +301,7 @@ internal fun Route.aktivitetspliktRoutes(
             post {
                 kunSkrivetilgang {
                     logger.info("Oppretter aktivitetsgrad for sakId=$sakId og oppgaveId=$oppgaveId")
-                    val aktivitetsgrad = call.receive<LagreAktivitetspliktAktivitetsgrad>()
+                    val aktivitetsgrad = call.receive<LagreAktivitetspliktAktivitetsgradMedUnntak>()
                     val aktivitetspliktVurdering =
                         inTransaction {
                             aktivitetspliktService.upsertAktivitetsgradForOppgave(
@@ -388,10 +388,10 @@ internal fun Route.aktivitetspliktRoutes(
         post("/aktivitetsgrad") {
             kunSkrivetilgang {
                 logger.info("Oppretter aktivitetsgrad for sakId=$sakId og behandlingId=$behandlingId")
-                val aktivitetsgrad = call.receive<LagreAktivitetspliktAktivitetsgrad>()
+                val aktivitetsgradOgUnntak = call.receive<LagreAktivitetspliktAktivitetsgradMedUnntak>()
                 inTransaction {
                     aktivitetspliktService.upsertAktivitetsgradForBehandling(
-                        aktivitetsgrad = aktivitetsgrad,
+                        aktivitetsgradOgUnntak = aktivitetsgradOgUnntak,
                         behandlingId = behandlingId,
                         sakId = sakId,
                         brukerTokenInfo = brukerTokenInfo,
